final_year_project/requirements.txt
streamlit==1.28.1
streamlit-option-menu==0.3.6
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
reportlab==4.0.4
Pillow==10.0.0
requests==2.31.0
joblib==1.3.1
matplotlib==3.7.1
seaborn==0.12.2
